Code update
完善了接口，添加了接口测试工具类TestUtil和测试数据生成类BeanProduct,日常接口的测试可以在InterfaceTest中进行，测
试方式参考示例

+++ src/main/java/com/blogadmin/blog/dao/BlogFileDao.java
@@ -11,11 +11,11 @@ import java.util.List;
  */
 public interface BlogFileDao {
 
-    int  logicRomove(Serializable id);
+    Long  logicRomove(Serializable id);
 
     Long updateEntity(BlogFile blogFile);
 
-    long saveEntity(BlogFile blogFile);
+    Long saveEntity(BlogFile blogFile);
 
     List<BlogFile> getAll();
